refactor(maps): move InteractiveMap to shared components
